requirements.txt
beautifulsoup4==4.9.1
pyowm==3.1.1
python-telegram-bot==20.7
Requests==2.31.0
